// web/src/app/generated/model/userList.ts
/**
 * Modul450 Library API
 * Simple library management system for renting books
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { User } from './user';


export interface UserList { 
    users: Array<User>;
}